feat: added theme preset support (#35)
* feat: theme presert support for end users

* chore: use generics to get the tailwind type from userland

* chore: resolve default tailwindconfig from userland

* feat: improved preset support
