+++ package.json
@@ -66,6 +66,7 @@
   },
   "dependencies": {
     "@renderlesskit/react": "^0.2.0",
+    "deepmerge": "^4.2.2",
     "reakit": "^1.3.4"
   },
   "devDependencies": {
