+++ preset.d.ts
@@ -1,0 +1,1 @@
+export function preset<T>(arg: T): T;
